Issue 5521 - UI - Update plugins for new split PAM and LDAP pass thru auth
Description:  Previously PAM and LDAP pass thru auth plugins were merged.  This change
separates them into their own plugins in the UI.

Also improved memory reporting in monitor tab.

relates: https://github.com/389ds/389-ds-base/issues/5521

Reviewed by: spichugi(Thanks!)
